Force invalidation of query set cache.
When a new object is created in Django Cache Machine, the original
behavior was to allow the query set to be “eventually consistent”
(https://github.com/jbalogh/django-cache-machine/issues/6).  Because
our admin needs to be immediately consistent, this fix forces
invalidation simply by invalidating the first item in the query set.